backend/python_service/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
transformers==4.48.3
torch>=2.1.0
pillow==10.4.0
python-multipart==0.0.9
sentencepiece==0.2.0
